requirements.txt
black==23.11.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
flake8==6.1.0
Flask==2.0.1
Flask-Cors==4.0.0
flask-mongoengine==1.0.0
Flask-Testing==0.8.1
flask-wtf==1.2.1
idna==3.6
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mock==5.1.0
mongoengine==0.27.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
pep8==1.7.1
platformdirs==4.0.0
pycodestyle==2.11.1
pyflakes==3.1.0
pymongo==4.6.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
tomli==2.0.1
typing-extensions==4.8.0
urllib3==2.1.0
Werkzeug==2.2.2
wtforms==3.1.1
zipp==3.17.0
